Move RoomBookingMapControl to its file
